+++ ui/coord/CalxCoordGraphCtrl.cpp
@@ -36,27 +36,27 @@ namespace CalXUI {
 		graphPanel->SetSizer(graphSizer);
 
 		this->expr = new wxTextCtrl(graphPanel, wxID_ANY, "x");
-        this->xmin = new wxSpinCtrlDouble(graphPanel, wxID_ANY, wxEmptyString, wxDefaultPosition,
-                                    wxDefaultSize, wxSP_ARROW_KEYS, INT_MIN, INT_MAX,
-									graphconf->getReal("x_from", -10.0),
-									graphconf->getReal("x_axis_step", 0.001));
-        this->xmax = new wxSpinCtrlDouble(graphPanel, wxID_ANY, wxEmptyString, wxDefaultPosition,
-                                    wxDefaultSize, wxSP_ARROW_KEYS, INT_MIN, INT_MAX,
-									graphconf->getReal("x_to", 10.0),
-									graphconf->getReal("x_axis_step", 0.001));
-        this->ymin = new wxSpinCtrlDouble(graphPanel, wxID_ANY, wxEmptyString, wxDefaultPosition,
-                                    wxDefaultSize, wxSP_ARROW_KEYS, INT_MIN, INT_MAX,
-									graphconf->getReal("y_from", -10.0),
-									graphconf->getReal("y_axis_step", 0.001));
-        this->ymax = new wxSpinCtrlDouble(graphPanel, wxID_ANY, wxEmptyString, wxDefaultPosition,
-                                    wxDefaultSize, wxSP_ARROW_KEYS, INT_MIN, INT_MAX,
-									graphconf->getReal("y_to", 10.0),
-									graphconf->getReal("y_axis_step", 0.001));
+		this->xmin = new wxSpinCtrlDouble(graphPanel, wxID_ANY, wxEmptyString, wxDefaultPosition,
+                                      wxDefaultSize, wxSP_ARROW_KEYS, INT_MIN, INT_MAX,
+									                    graphconf->getReal("x_from", -10.0),
+																		  graphconf->getReal("x_axis_step", 0.001));
+		this->xmax = new wxSpinCtrlDouble(graphPanel, wxID_ANY, wxEmptyString, wxDefaultPosition,
+                                      wxDefaultSize, wxSP_ARROW_KEYS, INT_MIN, INT_MAX,
+				                     					graphconf->getReal("x_to", 10.0),
+									                    graphconf->getReal("x_axis_step", 0.001));
+    this->ymin = new wxSpinCtrlDouble(graphPanel, wxID_ANY, wxEmptyString, wxDefaultPosition,
+                                      wxDefaultSize, wxSP_ARROW_KEYS, INT_MIN, INT_MAX,
+			                     						graphconf->getReal("y_from", -10.0),
+									                    graphconf->getReal("y_axis_step", 0.001));
+    this->ymax = new wxSpinCtrlDouble(graphPanel, wxID_ANY, wxEmptyString, wxDefaultPosition,
+                                      wxDefaultSize, wxSP_ARROW_KEYS, INT_MIN, INT_MAX,
+			                     						graphconf->getReal("y_to", 10.0),
+									                    graphconf->getReal("y_axis_step", 0.001));
 		this->step = new wxSpinCtrlDouble(graphPanel, wxID_ANY, wxEmptyString, wxDefaultPosition,
-                                    wxDefaultSize, wxSP_ARROW_KEYS, graphconf->getReal("step_step", 0.001), INT_MAX,
-									graphconf->getReal("step", 0.1),
-									graphconf->getReal("step_step", 0.001));
-        this->speed = new wxSpinCtrlDouble(graphPanel, wxID_ANY, wxEmptyString, wxDefaultPosition,
+                                      wxDefaultSize, wxSP_ARROW_KEYS, graphconf->getReal("step_step", 0.001), INT_MAX,
+		                     							graphconf->getReal("step", 0.1),
+									                    graphconf->getReal("step_step", 0.001));
+    this->speed = new wxSpinCtrlDouble(graphPanel, wxID_ANY, wxEmptyString, wxDefaultPosition,
                                      wxDefaultSize, wxSP_ARROW_KEYS, 0,
                                      wxGetApp().getSystemManager()->getConfiguration()->
                                         getEntry("ui")->getReal("unit_speed", 4000.0),
@@ -96,12 +96,12 @@ namespace CalXUI {
 		graphSizer->Add(buildButton);
 		graphSizer->Add(previewButton);
 
-		
+
 		ConfigManager *conf = wxGetApp().getSystemManager()->getConfiguration();
-		coord_point_t cen = {conf->getEntry("coords")->getInt("offset_x", 0),
-			conf->getEntry("coords")->getInt("offset_y", 0)};
-		coord_scale_t scl = {conf->getEntry("coords")->getInt("scale_x", 1),
-			conf->getEntry("coords")->getInt("scale_y", 1)};
+		coord_point_t cen = {static_cast<double>(conf->getEntry("coords")->getInt("offset_x", 0)),
+			static_cast<double>(conf->getEntry("coords")->getInt("offset_y", 0))};
+		coord_scale_t scl = {static_cast<double>(conf->getEntry("coords")->getInt("scale_x", 1)),
+			static_cast<double>(conf->getEntry("coords")->getInt("scale_y", 1))};
 		LinearCoordTranslator *basic = new LinearCoordTranslator(cen, scl);
 		this->trans = new ComplexCoordTranslator(basic);
 		this->translator = new CalxCoordFilterCtrl(this, wxID_ANY, this->trans);
